Rework how we do polling
